refactor: add CardSection component to SupportFundPage
